+++ lib/cloudflare-email.rb
@@ -3,5 +3,6 @@ require "cloudflare/email/error"
 require "cloudflare/email/response"
 require "cloudflare/email/client"
 require "cloudflare/email/secure_reply"
+require "cloudflare/email/secure_message_id"
 
 require "cloudflare/email/engine" if defined?(::Rails::Engine)